+++ src/potenziometer_manager.h
@@ -9,7 +9,7 @@ extern unsigned long T1;
 
 /// @brief Get the selected difficulty by reading the value from the potentiometer
 /// @return the selected difficulty
-int getDifficulty();
+uint8_t getDifficulty();
 
 /// @brief Initialize potentiometer's pin
 void potentiometerInit();
